requirements.txt
Pypubsub==4.0.3
twodict==1.2
wxPython==4.0.4
pyinstaller==3.6